refactor(match-service): change folder structure, file naming and remove unnecessary code
