+++ services/match_service/Dockerfile
@@ -2,12 +2,8 @@ FROM node:20 AS builder
 WORKDIR /app
 COPY package*.json ./
 RUN npm install
-COPY ./modules/ ./modules/
-COPY ./plugins ./plugins/
-COPY ./routes/ ./routes/
-COPY ./types ./types/
+COPY src/ ./src/
 COPY ./tsconfig.json ./
-COPY ./app.ts ./
 RUN npx tsc
 
 FROM node:20-slim
