+++ services/match_service/src/types/websocket.ts
@@ -1,0 +1,10 @@
+import {WebSocket} from "@fastify/websocket";
+
+export interface WsQuery {
+    playerJWT: string;
+}
+
+export interface MatchWebSocket extends WebSocket {
+    connectionId: string;
+    sessionId: string | null;
+}
