+++ services/match_service/tsconfig.json
@@ -9,6 +9,6 @@
     "strict": true,
     "skipLibCheck": true,
     "outDir": "./dist",
-    "rootDir": "./"
+    "rootDir": "./src"
   }
 }
