--- services/match_service/types/websocket.ts
@@ -1,7 +1,0 @@
-export interface WsParams {
-    gameId: string;
-}
-
-export interface WsQuery {
-    playerJWT: string;
-}
